新增 - 【招聘】获取用户角色列表 - 【招聘】查询面试官信息列表 - 【招聘】更新面试官信息 - 【招聘】获取面试速记明细 - 【应用信息】向管理员申请授权 - 【应用信息】查询租户授权状态
更新
- 通过地点 ID 批量获取地点信息 响应体
- 更新员工状态 请求体
- 更新员工状态 响应体
- 创建外部面试 请求体
- 创建外部面试 响应体
- 操作候选人入职 请求体

+++ src/Hire/GetHireV1QuestionnairesResponseDto.cs
@@ -269,12 +269,12 @@ public record GetHireV1QuestionnairesResponseDto
                 public string? LowestScoreEnDesc { get; set; }
 
                 /// <summary>
-                /// <para>评分分数</para>
+                /// <para>评分分数（整数）</para>
                 /// <para>必填：否</para>
                 /// <para>示例值：4</para>
                 /// </summary>
                 [JsonPropertyName("score_result")]
-                public float? ScoreResult { get; set; }
+                public int? ScoreResult { get; set; }
             }
 
             /// <summary>
@@ -297,7 +297,7 @@ public record GetHireV1QuestionnairesResponseDto
         public bool? HasAnswers { get; set; }
 
         /// <summary>
-        /// <para>满意度问卷更新时间</para>
+        /// <para>满意度问卷更新时间，毫秒时间戳</para>
         /// <para>必填：否</para>
         /// <para>示例值：1644412504000</para>
         /// </summary>
