+++ src/Hire/PostHireV1ApplicationsByApplicationIdTransferOnboardBodyDto.cs
@@ -14,7 +14,7 @@
 namespace FeishuNetSdk.Hire;
 /// <summary>
 /// 操作候选人入职 请求体
-/// <para>根据投递 ID 操作候选人入职并创建员工。投递须处于「待入职」阶段，可通过「转移阶段」接口变更投递状态。</para>
+/// <para>根据投递 ID 操作候选人入职并创建员工。后续可通过接口进一步获取入职信息，详情参考：[通过员工 ID 获取入职信息](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/employee/get)</para>
 /// <para>接口ID：6964286393804800028</para>
 /// <para>文档地址：https://open.feishu.cn/document/server-docs/hire-v1/candidate-management/delivery-process-management/onboard/transfer_onboard</para>
 /// <para>JSON地址：https://open.feishu.cn/document_portal/v1/document/get_detail?fullPath=%2fukTMukTMukTM%2fuMzM1YjLzMTN24yMzUjN%2fhire-v1%2fapplication%2ftransfer_onboard</para>
@@ -22,23 +22,23 @@ namespace FeishuNetSdk.Hire;
 public record PostHireV1ApplicationsByApplicationIdTransferOnboardBodyDto
 {
     /// <summary>
-    /// <para>实际入职时间</para>
+    /// <para>实际入职时间，毫秒时间戳</para>
     /// <para>必填：否</para>
     /// <para>示例值：1616428800000</para>
     /// </summary>
     [JsonPropertyName("actual_onboard_time")]
-    public int? ActualOnboardTime { get; set; }
+    public long? ActualOnboardTime { get; set; }
 
     /// <summary>
-    /// <para>预期转正时间</para>
+    /// <para>预期转正时间，毫秒时间戳</para>
     /// <para>必填：否</para>
     /// <para>示例值：1616428800000</para>
     /// </summary>
     [JsonPropertyName("expected_conversion_time")]
-    public int? ExpectedConversionTime { get; set; }
+    public long? ExpectedConversionTime { get; set; }
 
     /// <summary>
-    /// <para>招聘需求 ID</para>
+    /// <para>招聘需求 ID，可通过接口 [获取招聘需求列表](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/job_requirement/list) 获取。入职完成后招聘需求的「已入职」人数会加1</para>
     /// <para>必填：否</para>
     /// <para>示例值：6960663240925956402</para>
     /// </summary>
@@ -46,15 +46,16 @@ public record PostHireV1ApplicationsByApplicationIdTransferOnboardBodyDto
     public string? JobRequirementId { get; set; }
 
     /// <summary>
-    /// <para>操作人 UserID</para>
+    /// <para>操作人ID，与入参 `user_id_type` 类型一致</para>
     /// <para>必填：否</para>
-    /// <para>示例值：ou-xxx</para>
+    /// <para>示例值：7326856229396906012</para>
     /// </summary>
     [JsonPropertyName("operator_id")]
     public string? OperatorId { get; set; }
 
     /// <summary>
-    /// <para>候选人办公地点 ID ，枚举可通过接口「获取地址列表」获取，将用于候选人内推奖规则判断</para>
+    /// <para>候选人办公地点 ID，将用于候选人内推奖规则判断，数据源可通过接口[获取地址列表](https://open.feishu.cn/document/ukTMukTMukTM/uMzM1YjLzMTN24yMzUjN/hire-v1/location/list)</para>
+    /// <para>获取</para>
     /// <para>必填：否</para>
     /// <para>示例值：CT_2</para>
     /// </summary>
@@ -62,7 +63,7 @@ public record PostHireV1ApplicationsByApplicationIdTransferOnboardBodyDto
     public string? OnboardCityCode { get; set; }
 
     /// <summary>
-    /// <para>候选人入职部门 ID ，枚举可通过接口「获取部门信息列表」获取，将用于候选人内推奖规则判断</para>
+    /// <para>候选人入职部门 ID ，将用于候选人内推奖规则判断，可通过接口[搜索部门](https://open.feishu.cn/document/uAjLw4CM/ukTMukTMukTM/reference/contact-v3/department/search)获取，与入参 `department_id_type ` 类型一致</para>
     /// <para>必填：否</para>
     /// <para>示例值：6966123381141866028</para>
     /// </summary>
@@ -70,15 +71,15 @@ public record PostHireV1ApplicationsByApplicationIdTransferOnboardBodyDto
     public string? Department { get; set; }
 
     /// <summary>
-    /// <para>候选人直属上级 UserID ，将用于候选人内推奖规则判断</para>
+    /// <para>候选人直属上级 UserID ，将用于候选人内推奖规则判断，与入参 `user_id_type` 类型一致</para>
     /// <para>必填：否</para>
-    /// <para>示例值：ou-xxx</para>
+    /// <para>示例值：7326856229396906012</para>
     /// </summary>
     [JsonPropertyName("leader")]
     public string? Leader { get; set; }
 
     /// <summary>
-    /// <para>候选人序列 ID ，枚举可通过接口「获取职务分类列表」获取，将用于候选人内推奖规则判断</para>
+    /// <para>候选人序列 ID ，将用于候选人内推奖规则判断，与入参 `job_family_id_type ` 类型一致</para>
     /// <para>必填：否</para>
     /// <para>示例值：7006234385490345986</para>
     /// </summary>
@@ -86,7 +87,7 @@ public record PostHireV1ApplicationsByApplicationIdTransferOnboardBodyDto
     public string? Sequence { get; set; }
 
     /// <summary>
-    /// <para>候选人职级 ID ，枚举可通过接口「获取职级列表」获取，将用于候选人内推奖规则判断</para>
+    /// <para>候选人职级 ID ，将用于候选人内推奖规则判断，与入参 `job_level_id_type ` 类型一致</para>
     /// <para>必填：否</para>
     /// <para>示例值：6937934036379650311</para>
     /// </summary>
@@ -94,7 +95,7 @@ public record PostHireV1ApplicationsByApplicationIdTransferOnboardBodyDto
     public string? Level { get; set; }
 
     /// <summary>
-    /// <para>候选人入职人员类型 ID，可通过接口人力资源管理平台「获取员工类型列表」获取，将用于候选人内推奖规则判断</para>
+    /// <para>候选人入职人员类型 ID，将用于候选人内推奖规则判断，与入参 `employee_type_id_type ` 类型一致</para>
     /// <para>必填：否</para>
     /// <para>示例值：1</para>
     /// </summary>
